refactor: merged the app component tests into one file
